adicionado api key

+++ connectToHanaSpringBoot/src/main/java/com/example/demo/resources/UserService.java
@@ -3,10 +3,13 @@ package com.example.demo.resources;
 import com.example.demo.domain.User;
 import com.example.demo.repo.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.naming.AuthenticationException;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping(value = "/user")
@@ -15,13 +18,19 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
+    @Value("${apikey}")
+    private String apiKey;
+
     /**
      * retorna todos os users da base
      * @return lista de users
      */
     @RequestMapping(method = RequestMethod.GET)
-    public List<User> getAll(){
-        return userRepository.findAll();
+    public ResponseEntity getAll(@RequestHeader(value="Authorization") String key){
+        if (!Objects.equals(key, apiKey)){
+            return unauthorized();
+        }
+        return ResponseEntity.ok(userRepository.findAll());
     }
 
     /**
@@ -30,7 +39,10 @@ public class UserService {
      * @return response
      */
     @RequestMapping(method = RequestMethod.POST)
-    public ResponseEntity insert(@RequestBody User user){
+    public ResponseEntity insert(@RequestBody User user, @RequestHeader(value="Authorization") String key) {
+        if (!Objects.equals(key, apiKey)){
+            return unauthorized();
+        }
        userRepository.save(user);
        return ResponseEntity.ok().build();
     }
@@ -41,7 +53,10 @@ public class UserService {
      * @return respose
      */
     @RequestMapping(method = RequestMethod.PUT)
-    public ResponseEntity update(@RequestBody User user){
+    public ResponseEntity update(@RequestBody User user, @RequestHeader(value="Authorization") String key){
+        if (!Objects.equals(key, apiKey)){
+            return unauthorized();
+        }
         userRepository.save(user);
         return ResponseEntity.ok().build();
     }
@@ -52,8 +67,15 @@ public class UserService {
      * @return response
      */
     @RequestMapping(method = RequestMethod.DELETE)
-    public ResponseEntity insert(@RequestParam("id") int id){
+    public ResponseEntity insert(@RequestParam("id") int id, @RequestHeader(value="Authorization") String key) {
+        if (!Objects.equals(key, apiKey)){
+            return unauthorized();
+        }
         userRepository.deleteById(id);
         return ResponseEntity.ok().build();
+    }
+
+    private ResponseEntity unauthorized(){
+        return ResponseEntity.status(401).build();
     }
 }
